Front main

+++ Banco/src/main/java/com/reto/Banco/controller/clienteController.java
@@ -1,12 +1,19 @@
 package com.reto.Banco.controller;
 
 import java.time.LocalDate;
+import java.time.ZoneId;
+import java.time.format.DateTimeFormatter;
+import java.time.temporal.ChronoUnit;
+import java.util.Date;
 import java.util.List;
+import java.util.Optional;
 
+import org.hibernate.annotations.Any;
 import org.springframework.beans.factory.annotation.Autowired;
-
+import org.springframework.boot.autoconfigure.amqp.RabbitProperties.Retry;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.scheduling.Trigger;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -17,9 +24,16 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.reto.Banco.dto.GeneralResponse;
 import com.reto.Banco.entity.ClientTable;
+import com.reto.Banco.entity.ProductEntity;
 import com.reto.Banco.service.ClientService;
+import com.reto.Banco.service.ProductSevice;
+
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+
+import java.io.Console;
+import java.text.BreakIterator;
+import java.text.SimpleDateFormat;  
 
 @CrossOrigin(origins = "*")
 @RestController
@@ -28,19 +42,52 @@ public class clienteController {
 
     @Autowired
     ClientService clientService;
-
+    @Autowired 
+    ProductSevice productService;
     /**
      * @return
      */
     @GetMapping
-    public ResponseEntity<GeneralResponse<List<ClientTable>>> GetCliente() {
+    public ResponseEntity<GeneralResponse<List<ClientTable>>>  GetCliente() {
         GeneralResponse<List<ClientTable>> respuesta = new GeneralResponse<>();
 		List<ClientTable> datos = null;
 		String mensaje = null;	
 		HttpStatus estadoHttp = null;
 
         try {		
+            // clientService.GetAllCliente();
+            // System.out.println("Pointer");
             datos = clientService.GetAllCliente();
+           
+
+            respuesta.setDatos(datos);
+			respuesta.setMensaje(mensaje);
+			respuesta.setPeticionExitosa(true);	
+			estadoHttp = HttpStatus.OK;
+            
+        }
+        catch(Exception e)
+        {
+            mensaje = "500 Internal Server Error.Contact the administrator";			
+			respuesta.setMensaje(mensaje);
+			respuesta.setPeticionExitosa(false);
+			estadoHttp = HttpStatus.INTERNAL_SERVER_ERROR;
+        }
+        
+        return new ResponseEntity<>( respuesta,estadoHttp );
+        // return ;
+    }
+
+
+    @GetMapping("/{id}")
+    public ResponseEntity<GeneralResponse<Optional<ClientTable>>> GetClientById(@PathVariable("id") long id) {
+        GeneralResponse<Optional<ClientTable>> respuesta = new GeneralResponse<>();
+		Optional<ClientTable> datos = null;
+		String mensaje = null;	
+		HttpStatus estadoHttp = null;
+
+        try {		
+            datos = clientService.GetClienteById(id);
 
 
             respuesta.setDatos(datos);
@@ -65,24 +112,69 @@ public class clienteController {
         GeneralResponse<ClientTable> respuesta = new GeneralResponse<>();
         ClientTable datos = null;
 		String mensaje = null;	
-		HttpStatus estadoHttp = null;    
+		HttpStatus estadoHttp = null;   
+        ClientTable finalClient = null; 
 
         try{
+            // para testo en PostMan
+            // Date date1 = new SimpleDateFormat("dd/MM/yyyy").parse(client.getAuxBirthdate());          
+            // client.setBirthdate(date1);  
+            // System.out.println("Test: "+client.getAuxBirthdate());
+
+
+            // finalClient = new ClientTable(            client.getId_Type(),
+            //                                           client.getIdNum(),
+            //                                           client.getFisrtName(), 
+            //                                           client.getLastName(), 
+            //                                           client.getEmail(), 
+            //                                           date1,
+            //                                           client.getAuxBirthdate(),
+            //                                           "Admin",
+            //                                           LocalDate.now(),
+            //                                           date1 ,
+            //                                           "Admin");
+
+
+            // Date date1 = new SimpleDateFormat("dd/MM/yyyy").parse(client.getAuxBirthdate());          
+            // client.setBirthdate(date1);  
+            // System.out.println("Test: "+client.getAuxBirthdate());
+            // System.out.println(client.getIdNum());
+
+            finalClient = new ClientTable(            client.getId_Type(),
+                                                      client.getIdNum(),
+                                                      client.getFisrtName(), 
+                                                      client.getLastName(), 
+                                                      client.getEmail(), 
+                                                      client.getBirthdate(),
+                                                      client.getUserCreation(),
+                                                      LocalDate.now(),
+                                                      LocalDate.now(),
+                                                      client.getUserUpdate());                                
+
             
+                                                    
+
+                                                      
+
+        //  finalClient.setAuxBirthdate(client.getAuxBirthdate());
+            long years = ChronoUnit.YEARS.between( client.getBirthdate(),  LocalDate.now());
+           
+            System.out.println(years > 17);
             //restar edades
-            if(client.getAge() > 18 )
+            if (years > 17 )
             {
-                client.setDateCreation(LocalDate.now());
-                datos = clientService.CreateCliente(client);
+                // client.setDatecreation(LocalDate.now());
+                datos = clientService.CreateCliente(finalClient);
                 mensaje = "0 - Customer successfully created";
+                respuesta.setDatos(finalClient);
+                respuesta.setMensaje(mensaje);
             }else {
-                mensaje ="1 - Customer could not be create, the age ";
-				estadoHttp = HttpStatus.OK;
+                mensaje ="1 - Customer could not be create, the age";                
+				estadoHttp = HttpStatus.ALREADY_REPORTED;
             }
-
             
 
-            respuesta.setDatos(datos);
+
             respuesta.setMensaje(mensaje);
             respuesta.setPeticionExitosa(true);
             estadoHttp = HttpStatus.CREATED;
@@ -109,9 +201,30 @@ public class clienteController {
 		HttpStatus estadoHttp = null;    
 
         try{
-            client.setId(id);
-            datos = clientService.UpdateClient(client);
-            mensaje = "0 - Customer successfully created";
+            // client.setId(id);
+            
+
+
+            ClientTable  finalClient = new ClientTable(client.getId_Type(),
+                                                      client.getIdNum(),
+                                                      client.getFisrtName(), 
+                                                      client.getLastName(), 
+                                                      client.getEmail(), 
+                                                      client.getBirthdate(),
+                                                    //   client.getUserCreation(),
+                                                    "test",
+                                                    //   client.getDatecreation(),
+                                                    LocalDate.now(), 
+                                                      LocalDate.now(),
+                                                      //client.getUserUpdate()
+                                                      "Admin"
+                                                      ); 
+            
+                        System.out.println("Test");
+            finalClient.setId(id);
+
+            datos = clientService.UpdateClient(finalClient);
+            mensaje = "0 - Customer successfully Update";
 
             respuesta.setDatos(datos);
             respuesta.setMensaje(mensaje);
@@ -135,17 +248,50 @@ public class clienteController {
         GeneralResponse<Long> respuesta = new GeneralResponse<>();        
 		String mensaje = null;	
 		HttpStatus estadoHttp = null;    
-
+        List<ProductEntity> Products; 
+        boolean trigger = false;
+        
         try{
-            // client.setId(id);
-            clientService.delete(id);
+            
+            Products =  productService.findByclienteId(id);
+            // System.out.println(Products.size());
+
+
+            
+            
+            for(int i = 0 ; i < Products.size(); i++ )
+            {
+                if(!Products.get(i).getState().toString().equals("cancelled") )
+                {
+                    // System.out.println(!Products.get(i).getEstado().equalsIgnoreCase("cancelled"));
+                    trigger = true;
+                    break;
+                }
+            }
+            
+            System.out.println("Trigger Value: "+ trigger  );
+            if( trigger == false)
+            {
+                clientService.delete(id);
+
+                for(int i = 0 ; i < Products.size(); i++ )
+                {
+                    productService.deleteById(Products.get(i).getId());
+                }
+                
+                mensaje = "0 - Client successfully delete";
+                respuesta.setPeticionExitosa(true);
+                estadoHttp = HttpStatus.ACCEPTED;
+            }else 
+            {
+                mensaje = "1 - Customer don't be  delete, exist active product.";
+                respuesta.setPeticionExitosa(true);
+                estadoHttp = HttpStatus.ALREADY_REPORTED;
+            }
            
-            mensaje = "0 - Customer successfully delete";
 
             respuesta.setDatos(id);
             respuesta.setMensaje(mensaje);
-            respuesta.setPeticionExitosa(true);
-            estadoHttp = HttpStatus.CREATED;
         }catch(Exception e){
 
             estadoHttp = HttpStatus.INTERNAL_SERVER_ERROR;
